Add missing headers and fix warnings

+++ src/util.h
@@ -6,6 +6,7 @@
 #include <SDL3/SDL.h>
 #include <algorithm>
 #include <random>
+#include <utility>
 
 struct Random {
 private:
@@ -61,7 +62,7 @@ auto inline viscosity(const cell_t &cell) {
  */
 auto inline density_le_chance(const cell_t &a, const cell_t &b, Random &rng) {
     auto diff = density(b) - density(a);
-    return diff && rng.gen_real() < diff;
+    return static_cast<int>(diff) && rng.gen_real() < diff;
 }
 
 SDL_AppResult SDL_Fail();
